Split files and add barrels when necessary
