+++ frontend/src/metabase/visualizations/components/ChartSettings/DashboardChartSettings/index.ts
@@ -1,0 +1,1 @@
+export * from "./DashboardChartSettings";
