+++ frontend/src/metabase/visualizations/components/ChartSettings/QuestionChartSettings/index.ts
@@ -1,0 +1,1 @@
+export * from "./QuestionChartSettings";
